requirements.txt
BeautifulSoup==3.2.1
PyYAML==3.11
argparse==1.2.1
args==0.1.0
bottle==0.12.9
clint==0.3.3
docopt==0.6.2
internetarchive==0.9.8
jsonpatch==0.4
peewee==2.8.0
python-dateutil==2.5.0
requests==2.7.0
six==1.10.0
wsgiref==0.1.2
